refactor identity manager
